requirements.txt
lxml==4.4.2
python-dotenv==0.10.3
six==1.13.0
SQLAlchemy==1.3.12
SQLAlchemy-Utils==0.36.1
tqdm==4.41.0
uni2ascii-janin==1.0.1
Unidecode==1.1.1
